+++ src/cli.js
@@ -113,9 +113,9 @@ Project type-specific commands:
     Create ES5, ES6 modules and UMD builds for a React component.
 
     Options:
-      ${opt('--copy-files')}    copy files which won't be transpiled by Babel (e.g. CSS)
-      ${opt('--no-demo')}       don't build the demo app, if present
-      ${opt('--no-proptypes')}  don't wrap propTypes with an environment check
+      ${opt('--copy-files')}        copy files which won't be transpiled by Babel (e.g. CSS)
+      ${opt('--no-demo')}           don't build the demo app, if there is one
+      ${opt('--[keep-]proptypes')}  keep component propTypes in production builds
 
   ${cmd('nwb build-preact-app')} ${opt('[entry] [dist_dir]')}
     Build a Preact app from ${opt('entry')} to ${opt('dist_dir')}.